Continuous_Integration_Deployment/CI_CD_Dockers/model_training_serving_workflow/requirements.txt
pandas
scikit-learn
pyyaml
joblib
mlflow
fastapi
uvicorn
boto3